refactor(npm): prepartion for yarn catalogs integratoin
